% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_edgelist.R
\name{as_edgelist}
\alias{as_edgelist}
\title{Restructure interactions as an edgelist}
\usage{
as_edgelist(
  interactions,
  min_secs_elapsed = 4,
  max_secs_elapsed = Inf,
  min_proportion_barn = 0,
  max_proportion_barn = 1
)
}
\arguments{
\item{interactions}{Data frame of interactions between actors.}

\item{min_secs_elapsed}{Minimum number of seconds elapsed for an interaction to count.}

\item{max_secs_elapsed}{Maximum number of seconds elapsed for an interaction to count.}

\item{min_proportion_barn}{Minimum proportion of meetings taking place in the barn for an interaction to count.}

\item{max_proportion_barn}{Maximum proportion of meetings taking place in the barn for an interaction to count.}
}
\value{
A data frame.
}
\description{
Restructure interactions as an edgelist
}
\details{
Interactions are filtered according to the seconds elapsed in the
interaction and proportion of meetings taking place in the barn. Once a subset
of the interactions of interest has been obtained, the resulting data frame is
summarized into an edgelist by counting the number of resulting interactions
between each pair of actors.
}
