% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds}
\alias{birds}
\title{Tags and pairs for barn swallows.}
\format{
A data frame with 34 rows and 7 variables:
\describe{
\item{year}{year of study}
\item{nest_id}{nest ID}
\item{nest}{nest}
\item{bird_id}{bird ID}
\item{tag}{tag number}
\item{sex}{sex}
\item{sex}{tag number for pair}
}
}
\source{
\url{https://docs.google.com/document/d/15JlJuqI2GuHtTGzNFy0lefYcYhPoZQKN/edit?usp=sharing&ouid=112915548451337201344&rtpof=true&sd=true}
}
\usage{
birds
}
\description{
A dataset containing the tags, sex, and pairs for each of the 34 barn swallows
from 2021 and beyond.
}
\keyword{datasets}
