% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_barn_meetings.R
\name{classify_barn_meetings}
\alias{classify_barn_meetings}
\title{Append whether the sender and receiver nodes were registered by the base station.}
\usage{
classify_barn_meetings(mnData, irData, time_eps = 2, barn_rule = "or")
}
\arguments{
\item{mnData}{Mobile node data frame.}

\item{irData}{In range data frame.}

\item{time_eps}{A numeric for the maximum number of seconds the in range timestamp
can deviate from the mobile node timestamp while claiming a bird is in the barn.}

\item{barn_rule}{Rule for determining whether the meeting took place in barn.
Option \code{and} signifies both birds need be in the barn; whereas option \code{or}
signifies at least one bird needs to be in the barn.}
}
\value{
A data frame.
}
\description{
Append whether the sender and receiver nodes were registered by the base station.
}
\details{
If the sender/receiver node is registered by a base station no more
than \code{time_eps} seconds before or after the time of the meeting, then the
sender/receiver is deemed near the base station at the time of the meeting.
The meeting is deemed near the base station according to the \code{barn_rule}.
}
\examples{

if(FALSE) {
mnData <- read_mobilenode(path = 'dulogs/2022', pattern = 'MN_DATA', year = 2022)
irData <- read_inrange(path = 'dulogs/2022', pattern = 'MN_RANGE', year = 2022)
classify_barn_meetings(mnData = mnData, irData = irData, time_eps = 2, barn_rule = 'or')
}
}
\keyword{internal}
