% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_reciprocal_meetings.R
\name{classify_reciprocal_meetings}
\alias{classify_reciprocal_meetings}
\title{Append whether meeting is deemed reciprocal.}
\usage{
classify_reciprocal_meetings(dulogs, time_eps = 2)
}
\arguments{
\item{time_eps}{A numeric for the maximum number of seconds the timestamp's can
deviate while classifying a meeting as reciprocal (see details).}

\item{pdulogs}{Data frame of pre-processed dulog data.}
}
\value{
A data frame.
}
\description{
Append whether meeting is deemed reciprocal.
}
\details{
At a given meeting time, a receiver registers a signal from a sender.
This meeting is deemed reciprocal if the sender also registers a signal from
the sender no more than \code{time_eps} seconds before or after the meeting time.
}
\examples{
if(FALSE) {
dulogs <- read_dulogs(
paths = c(mobilenode = 'dulogs/2022', inrange = 'dulogs/2022'),
patterns = c(mobilenode = 'MN_DATA', inrange = 'MN_RANGE'),
year = 2022,
time_eps = 2,
barn_rule = 'or')
)
pdulogs <- preprocess_dulogs(dulogs)
classify_reciprocal_meetings(pdulogs, time_eps = 2)
}
}
\keyword{internal}
