% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_interactions.R
\name{construct_interactions}
\alias{construct_interactions}
\title{Construct data frame of interactions.}
\usage{
construct_interactions(
  dulogs,
  time_gap = 8,
  directed = FALSE,
  loops = FALSE,
  log = NULL
)
}
\arguments{
\item{time_gap}{A numeric for the maximum number of seconds between meetings
before a new interaction is defined.}

\item{directed}{Logical indicating whether to treat as directed.}

\item{loops}{Logical indicating whether to include self loops.}

\item{pdulogs}{Data frame of pre-processed dulog data.}
}
\value{
A data frame.
}
\description{
Construct data frame of interactions.
}
