% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_dulogs.R
\name{preprocess_dulogs}
\alias{preprocess_dulogs}
\title{Preprocess dulogs. See details.}
\usage{
preprocess_dulogs(
  dulogs,
  min_rssi = -Inf,
  max_rssi = 15,
  start_time = lubridate::make_datetime(year = 0),
  end_time = lubridate::make_datetime(year = 9999),
  only_barn = FALSE,
  log = NULL
)
}
\arguments{
\item{dulogs}{Data frame of dulog data.}

\item{min_rssi}{Minimum RSSI to consider.}

\item{max_rssi}{Maximum RSSI to consider.}

\item{start_time}{Earliest date and time to consider.}

\item{end_time}{Latest date and time to consider.}

\item{only_barn}{Logical indicating whether to keep only meetings in the barn.}

\item{log}{File path and name for preprocessing log.}
}
\value{
A data frame.
}
\description{
Preprocess dulogs. See details.
}
\details{
Meetings must satisfy the time frame of study including start and end dates
and daily monitoring times. These yearly constraints are provided in \code{data-raw/timeframe.R}.
Only birds under study are considered; these birds are documented in \code{data-raw/birds.R}.
Any meeting involving a phantom tag is subsequently removed. Proximity of birds
is proxied by signal strength. Some meetings are perfect duplicates (i.e.
having the same sender, receiver, meeting time, and signal strength) and are
subsequently removed. Other meetings are near duplicates (i.e. having the same
sender, receiver, and meeting time). In this case, the log registering the weakest
signal strength is preserved. Meetings with a signal strength weaker than the
provided upperbound are removed. If requested, only meetings in the barn are
considered, and finally, all meetings must take place before the start and end
times specified.
}
\examples{
if(FALSE) {
dulogs <- read_dulogs(
paths = c(mobilenode = 'dulogs/2022', inrange = 'dulogs/2022'),
patterns = c(mobilenode = 'MN_DATA', inrange = 'MN_RANGE'),
year = 2022,
time_eps = 2,
barn_rule = 'or'
)
preprocess_dulogs(
dulogs,
min_rssi = -Inf,
max_rssi = 15,
end_time = '0000-01-01 UTC',
start_time = '9999-01-01 UTC',
only_barn = FALSE
)
}
}
