% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dulogs.R
\name{read_dulogs}
\alias{read_dulogs}
\title{Read and combine dulog files. See details.}
\usage{
read_dulogs(paths, patterns, year, time_eps = 2, barn_rule = "or", log = NULL)
}
\arguments{
\item{paths}{A named character vector with two named elements: \code{mobilenode} and
\code{inrange}, each specifying the directory containing the mobile node and node
range data, respectively.}

\item{patterns}{A named character vector with two named elements: \code{mobilenode} and
\code{inrange}, each specifying a regular expression. Only file names within the
directory which match the regular expression will be returned.}

\item{year}{A character string, YYYY, indicating the year under study.}

\item{time_eps}{A numeric for the maximum number of seconds the in range timestamp
can deviate from the mobile node timestamp while claiming a bird is in the barn.}

\item{barn_rule}{Rule for determining whether the meeting took place in barn.
Option \code{and} signifies both birds need be in the barn; whereas option \code{or}
signifies at least one bird needs to be in the barn.}

\item{log}{File path and name for preprocessing log.}
}
\value{
A data frame.
}
\description{
Read and combine dulog files. See details.
}
\details{
Combines dulog files from a given year, including those data from the
mobile-node (named \code{mobilenode}) and mobile range (named \code{inrange}) csv files.
The user provides a named character vector specifying the directories, and a
named character vector of regular expression patterns for which only matching
files will be returned. These files are combined when possible into a data frame
with tidy column names and proper data types. The in range data are used as a
proxy from whether meetings between actors occured in a barn (where the base
stations are located).
}
\examples{
if(FALSE) {
read_dulogs(
paths = c(mobilenode = 'dulogs/2022', inrange = 'dulogs/2022'),
patterns = c(mobilenode = 'MN_DATA', inrange = 'MN_RANGE'),
year = 2022,
time_eps = 2,
barn_rule = 'or'
)
}
}
