% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_inrange.R
\name{read_inrange}
\alias{read_inrange}
\title{Read and combine in range data files. See details.}
\usage{
read_inrange(path, pattern, year)
}
\arguments{
\item{path}{A character vector of full path name.}

\item{pattern}{A regular expression. Only file names which match the regular
expression will be returned.}

\item{year}{A character string, YYYY, indicating the year under study.}
}
\value{
A data frame.
}
\description{
Read and combine in range data files. See details.
}
\details{
Combines in range data files from a given year. The user provides
the directory containing the in range \code{csv} files and a regex pattern to
retrieve the in range files. These files are combined when possible
into a data frame with tidy column names and proper data types.
}
\examples{
if(FALSE) {
irData <- read_inrange(path = 'dulogs/2022', pattern = 'MN_RANGE', year = 2022)
}
}
\keyword{internal}
