% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mobilenode.R
\name{read_mobilenode}
\alias{read_mobilenode}
\title{Read and combine mobile-node data files. See details.}
\usage{
read_mobilenode(path, pattern, year)
}
\arguments{
\item{path}{A character vector of full path name}

\item{pattern}{A regular expression. Only file names which match the regular
expression will be returned.}

\item{year}{A character string, YYYY, indicating the year under study.}
}
\value{
A data frame.
}
\description{
Read and combine mobile-node data files. See details.
}
\details{
Combines mobile-node data files from a given year. The user provides
the directory containing the mobile-node \code{csv} files and a regex pattern to
retrieve the mobile node data files. These files are combined when possible
into a data frame with tidy column names and proper data types.
}
\examples{
if(FALSE) {
read_mobilenode(path = 'dulogs/2022', pattern = 'MN_DATA', year = 2022)
}
}
\keyword{internal}
