% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeframe}
\alias{timeframe}
\title{Time frame of barnswallow study.}
\format{
A data frame with 1 row and 3 variables:
\describe{
\item{year}{year of study}
\item{start}{start date and time of data collection, format: HH:MM:SS YYYY-MM-DD UTC}
\item{end}{end date and time of data collection, format: HH:MM:SS YYYY-MM-DD UTC}
}
}
\source{
\url{https://docs.google.com/document/d/15JlJuqI2GuHtTGzNFy0lefYcYhPoZQKN/edit?usp=sharing&ouid=112915548451337201344&rtpof=true&sd=true}
}
\usage{
timeframe
}
\description{
A dataset containing the start and end dates for the 2021 barn swallow study
and beyond. The start and end times dictate when data are collected from the
swallows each day under study.
}
\keyword{datasets}
