% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DulogInteractions-package.R
\docType{package}
\name{DulogInteractions-package}
\alias{DulogInteractions}
\alias{DulogInteractions-package}
\title{DulogInteractions: Compiles data from dulog tags.}
\description{
Dulog tags collect relational data between actors wearing the tags. This package compiles dulog data into meaningul interactions.
}
\author{
\strong{Maintainer}: Connor Gibbs \email{connor.gibbs@colostate.edu} (\href{https://orcid.org/0000-0002-8852-5594}{ORCID})

}
\keyword{internal}
